+++ Example/MemfaultCloud_DemoApp/AppDelegate.swift
@@ -1,5 +1,5 @@
 ///
-/// Copyright (c) 2020-Present Memfault, Inc.
+/// Copyright (c) Memfault, Inc.
 /// See LICENSE for details
 ///
 
@@ -10,12 +10,12 @@ import MemfaultCloud
 class AppDelegate: UIResponder, UIApplicationDelegate {
     var window: UIWindow?
 
-    lazy var api: MemfaultApi = MemfaultApi(configuration: [
-        kMFLTProjectKey: "<YOUR_PROJECT_KEY_HERE>",
-    ])
-
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
         gMFLTLogLevel = .debug
+
+        MemfaultApi.configureSharedApi([
+            kMFLTProjectKey: "<YOUR_PROJECT_KEY_HERE>",
+        ])
         return true
     }
 }